Pass buildConfig when getting path to last build output
Pass the whole buildConfig object when getting the path to the last build output. This required changes in debug services and TestExecutionService.
Also change the args of cleanProject method - the `options` passed to `gradle` are now determined in the method itself instead of passing them from the caller.
This way we'll not require "android-23" (the default one set in build.gradle) when we miss to pass the required options.