Track the amount and device type of builds and livesync operations
In order to unsderstand the workflow of our users, we need to track how many builds they are executing, how many livesync operations, etc. We also need to know if the operation is for device or emulator (iOSSimulator).
This will allow us to focus the development and provide better features in the future.