Track the amount and device type of builds and livesync operations (#2690)
* Track the amount and device type of builds and livesync operations

In order to unsderstand the workflow of our users, we need to track how many builds they are executing, how many livesync operations, etc. We also need to know if the operation is for device or emulator (iOSSimulator).
This will allow us to focus the development and provide better features in the future.

* Track information for deploy and for device OS

Track information for deploy operations. Introduce new method in platformService, that tracks the action.
Track the device OS version as well.

* Track Android only for build operations

Track only Android platform for build operation. We do not need to track Android.emulator and Android.device for build operations as they are both the same actions.