Move logic for getting PID of application from iOS Sim to common
Move the logic for getting PID of application started on iOS Simulator to mobile-cli-lib.
Add new helper method to get the process ID of an application from iOS Simulator Logs. Whenever we start an application on iOS Simulator, in its logs we can find the PID of the process in the following format:
```
<app id>: <app id>: <PID>
```
Add unit tests for the method.