Fix debug android command (#2713)
* Fix Android debugging

When executing `tns debug android` for slow emulators/devices we have logic to wait several seconds until debugger is attached.
However we've missed to `await` the `sleep` method and in fact we are not waiting at all. Add the missing await, which fixes the `tns debug android` command.

* Remove unused code

A while ago, we had a logic to find out which node_modules should be processed. However it is not used anymore (removed in https://github.com/NativeScript/nativescript-cli/pull/2152/ ).
So remove the code - we do not need it anymore.

* Fix type