Fix `tns debug ios --justlaunch`
When `--justlaunch` flag is passed, CLI automatically disposes the `ios-device-lib` after execution of the command. However this closes all opened device sockets and breaks the debug command.
So whenever you try `tns debug ios --justlaunch [--chrome]`, the debug application (NativeScript Inspector) or Chrome DevTools is unable to connect to device socket and fails with error.
Fix this by keeping the `ios-device-lib` alive whenever we try debug for ios. This way the debug application will be able to connect to the device.
NOTE: `tns debug ios --justlaunch [--chrome]` cannot release the terminal as it has to keep the ios-device-lib alive.