Fix always reinstalling apps during livesync
In case the app is not installed on a device when `tns run <platform>` is started, each change reinstalls the application.
The problem is in the "isApplicationInstalled" check which was relying on the result of the first execution (when the app has not been installed).
Fix this by always checking if the app is installed.