Fix execution of livesync hooks when there are no node_modules
In case `node_modules` are deleted from the project and `tns run <platform>` or `tns debug <platform>` is executed, the livesync hook fails as the package that's required in it is not installed.
In order to fix this, extract the logic of the liveSync method in a separate method and decorate it with the hook. In the public method make sure all dependencies are installed and call the private method.
This way the hook will be executed correctly.