Allow prepare for Android without ANDROID_HOME (#2904)
* Allow prepare for Android without ANDROID_HOME

Allow preparing Android project without the requirement to have ANDROID_HOME set.
The change is in the required services, which should work without ANDROID_HOME in some cases.
However the commands should not allow this, so move the validation in the commands.

* Fix unable to uninstall application on macOS

Update ios-device-lib where a fix for uninstalling application on macOS was not working.
Update common lib where the following fixes are applied:
* Fix `tns devices` not showing attached iOS devices.
* Do not call uninstall of application when it is not installed - this produces some confusing warnings.

* Fix execution of livesync hooks when there are no node_modules

In case `node_modules` are deleted from the project and `tns run <platform>` or `tns debug <platform>` is executed, the livesync hook fails as the package that's required in it is not installed.
In order to fix this, extract the logic of the liveSync method in a separate method and decorate it with the hook. In the public method make sure all dependencies are installed and call the private method.
This way the hook will be executed correctly.