Fix merge of options during `tns run`
The `$options` object has the required information for all passed `--` args, but all of them are in `get` methods.
Calling `merge` does not get functions (like `get`ers), so we do not receive expected options (like `--release, --keyStorePath`, etc.) in the merged object.
In order to fix this just use the `this.$options.argv` which contains the information we need.