Fix asking for user email on postinstall
On postinstall CLI asks the users if they want to subscribe for NativeScript's newsletter. This check has been broken in 3.0.0 because of incorrect await.
In order to add tests for this functionality, introduce a new service that handles the check and sending of information about the user's mail.
Add the following tests:
 - ensure postinstall.js calls correct CLI command
 - ensure the post-install-cli command calls the new service
 - ensure the new service works correctly