Fix plugins prompts on postinstall
In case some plugins (like `nativescript-plugin-firebase`) try to prompt the user on postinstall, users are unable to answer the question when `tns plugin add <name>` is used.
The reason is that the stdin of current process is not passed to the stdin of the `npm install` process and the user has no way to input the result.
Fix this by using "inherit" of current stdio when the console is interactive.