Ensure symlinks from plugins are copied from realpath to platforms dir
In case a plugin is symlinked or it has symlink files/directories, we do not want to copy the symlink to platforms dir, but the real files. This is required as the symlink may be relative to the current location and copying it makes it invalid.
So pass `-L` option to `shelljs.cp`, so the symlinked files will be followed and the real ones will be copied.

This also fixes unit tests with npm5.