Add official support for Node.js 8 (#2979)
* Add official support for Node.js 8

Add npm-shrinkwrap.json that is required for building CLI's package with npm 5. By default npm 5 creates package-lock.json, but it cannot be published. So in case we use `package-lock.json`, we'll always have exact dependency tree when using the repository, but end-users will have different dependency tree. So instead of using package-lock.json we'll use npm-shrinkwrap.json that can be published in npm.

Update engine's tag in package.json, so in case Node.js 8 is used, we'll no longer show warning to the users.

* Ensure symlinks from plugins are copied from realpath to platforms dir

In case a plugin is symlinked or it has symlink files/directories, we do not want to copy the symlink to platforms dir, but the real files. This is required as the symlink may be relative to the current location and copying it makes it invalid.
So pass `-L` option to `shelljs.cp`, so the symlinked files will be followed and the real ones will be copied.

This also fixes unit tests with npm5.