Fix debug when using CLI as library
When CLI is used as a library, the method for debug returns incorrect result - previously the return type of the method we are calling was `string[]`, we've changed it to `string` but forgot to remove the `_.first(result)`.
So now the debug returns incorrect URL. Remove the `_.first` and return the result directly.