Make `tns run` without platform working
Introduce new method in livesync-command-helper to get platforms on which to execute action.
Rename a method in the helper as its purpose is to execute livesync, not to get params for livesync operation.
Pass correct args to detectCurrentlyAttachedDevices in debug command.
Make `tns run` to fail in case there's no running emulators and no devices attached as we are unable to determine for which platform to stat emulator.