Fix parsing of CLI's output
The sysInfo spawns NativeScript CLI's commands to get CLI's version and the version of `nativescript-cloud`. However, in some cases, CLI's output contains warnings that Node.js version is not supported. These warnings contain version of Node.js and current code decides that this is the version of CLI (and nativescript-cloud). Fix the parsing by using different regular expression - in CLI's output the version is always on a new line, so rely on this.
Add unit tests for this scenario and for getting the version of nativescript-cloud.