Fix `tns debug <platform> --start` inconsistent behavior
`tns debug <platform>` will check all available devices and will prompt the user to select device for debugging. In case the terminal is not interactive, it will select the emulator/device with highest API level.
However, adding `--start` causes failure as the described logic is not used. So ensure we have this behavior even when `--start` is used.